backend/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Proj"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.10",
    "llama-index-embeddings-gemini>=0.3.1",
    "llama-index-graph-stores-neo4j>=0.4.5",
    "llama-index-llms-gemini>=0.4.3",
    "llama-index-readers-wikipedia>=0.3.0",
    "wikipedia>=1.4.0",
    "python-dotenv>=1.0.1",
    "nest-asyncio>=1.6.0",
    "fastapi[standard]",
    "pydantic>=2.10.4",
    "pyvis>=0.3.2",
    "neo4j>=5.27.0",
    "ipysigma>=0.24.5",
    "starlette>=0.45.3",
]

# [project.scripts]
# hello = "hello_world:hello"
#
# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

# [dependency-groups]
# dev = [
#     "ipykernel>=6.29.5",
#     "ruff>=0.6.7",
# ]
